requirements.txt
ytmusicapi
python-slugify